+++ recipes/default.rb
@@ -7,4 +7,4 @@
 include_recipe 'chef_work_environment::source_config'
 include_recipe 'chef_work_environment::packages'
 include_recipe 'chef_work_environment::chef_workstation'
-include_recipe 'chef_work_environment::gitlab_runner'
+include_recipe 'chef_work_environment::repin'