+++ CHANGELOG.md
@@ -3,9 +3,23 @@
 This file is used to list changes made in each version of the
 chef_work_environment cookbook.
 
+## v0.2.0 - 10.03.2021
+
+- Add `repin` recipe to override chef bundled gem versions
+- Add example policyfile for use
+- Add configurability for Omnitruck, Chef Server/Supermarket and Rubygems
+- Add more Chef environment variables
+- Add global Chef configuration file
+- Add `vault` recipe
+- Add tagging to `gitlab_runner`
+- Remove `gitlab_runner` from `default` recipe
+- Update `gitlab_runner` recipe to v13.10.0
+- Fix kitchen.yml
+- Fix tests
+- Fix chef_workstation version to work on Debian/Ubuntu as well
+- Fix config file parameter on `gitlab_runner`
+- Fix executor type on `gitlab_runner`
+
 ## v0.1.0 - 17.11.2020
 
 Initial release.
-
-- change 0
-- change 1